+++ static/components/ui.js
@@ -8,3 +8,7 @@ import "./u-preview.js";
 import "./u-form.js";
 import "./u-pagepreview.js";
 import "./u-confirm.js";
+
+if (!("anchorName" in document.documentElement.style)) {
+  import("https://unpkg.com/@oddbird/css-anchor-positioning");
+}