+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.3.13] - Unreleased
+## [0.4.0] - Unreleased
+### Changed
+- More compact UI.
+- `f-object` elements are closed by default. Use `attributes.open = true` to change it.
+- Removed `x` button of popovers.
+
 ### Fixed
 - `file` field when the value is an URL.
 - Removed the numbers in the choose-list items.
@@ -218,7 +223,7 @@ First version
 [#9]: https://github.com/lumeland/cms/issues/9
 [#10]: https://github.com/lumeland/cms/issues/10
 
-[0.3.13]: https://github.com/lumeland/cms/compare/v0.3.12...HEAD
+[0.4.0]: https://github.com/lumeland/cms/compare/v0.3.12...HEAD
 [0.3.12]: https://github.com/lumeland/cms/compare/v0.3.11...v0.3.12
 [0.3.11]: https://github.com/lumeland/cms/compare/v0.3.10...v0.3.11
 [0.3.10]: https://github.com/lumeland/cms/compare/v0.3.9...v0.3.10